pyproject.toml
[tool.poetry]
name = "mytodo"
version = "0.1.0"
description = ""
authors = ["Алексей <Starobogatov.A@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.2"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqladmin = {extras = ["full"], version = "^0.16.1"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
aioredis = "^2.0.1"
pytest-asyncio = "^0.23.6"
bcrypt = "4.0.1"
pytest-cov = "^5.0.0"
celery = {extras = ["redis"], version = "^5.4.0"}
pytz = "^2024.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
